Update lob.py
LimitOrderBook.remove(order) now uses the order it matches in the order list with the order_id of the order argument for the remove order logic rather than the order argument itself.

During an update it's possible that we know we want to remove an order from the book (size=0) but not know if that order is a Ask or a Bid. For example, a valid BitFinex update from the websocket api may look like this: [10699330795, 8094.4, 0] telling us that we should remove that order from the order book, but not telling us if that order was a Bid or an Ask.